backend/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
openai==1.57.2
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
google-generativeai==0.8.5
google-cloud-texttospeech==2.17.2
python-dotenv==1.1.1
slowapi==0.1.9
sqlalchemy==2.0.43
bcrypt==5.0.0
passlib==1.7.4
python-jose[cryptography]==3.5.0
email-validator==2.2.0 
